requirements.txt
# VYRA Local Repository — Python dependencies
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
aiofiles>=23.2.1
pyyaml>=6.0
jsonschema>=4.22.0
